@@ -3,9 +3,10 @@
   "description": "microservice utils based on amqp transport layer",
   "main": "./lib/index.js",
   "scripts": {
-    "compile": "./node_modules/.bin/babel -d ./lib ./src",
+    "compile": "babel -d ./lib ./src",
     "prepublish": "npm run compile",
-    "test": "./node_modules/.bin/mocha",
+    "lint": "eslint ./src",
+    "test": "npm run lint && mocha",
     "bench": "npm run compile && node ./bench/roundtrip.js",
     "semantic-release": "semantic-release pre && npm publish && semantic-release post"
   },
@@ -20,25 +21,23 @@
   },
   "homepage": "https://github.com/makeomatic/ms-amqp-transport#readme",
   "devDependencies": {
-    "babel-cli": "^6.4.5",
-    "babel-eslint": "^5.0.0-beta6",
+    "babel-cli": "^6.4.0",
+    "babel-eslint": "^6.0.0",
     "babel-plugin-transform-class-properties": "^6.4.0",
-    "babel-plugin-transform-es2015-destructuring": "^6.4.0",
-    "babel-plugin-transform-es2015-parameters": "^6.4.5",
-    "babel-plugin-transform-es2015-shorthand-properties": "^6.3.13",
-    "babel-plugin-transform-es2015-spread": "^6.4.0",
+    "babel-plugin-transform-es2015-spread": "^6.3.13",
     "babel-plugin-transform-object-rest-spread": "^6.3.13",
     "babel-plugin-transform-strict-mode": "^6.3.13",
     "babel-register": "^6.4.3",
     "benchmark": "^2.0.0",
     "chai": "^3.4.1",
     "cz-conventional-changelog": "^1.1.5",
-    "eslint": "^1.10.3",
-    "eslint-config-airbnb": "^4.0.0",
-    "eslint-plugin-mocha": "^1.1.0",
+    "eslint": "^2.9.0",
+    "eslint-config-airbnb-base": "^3.0.1",
+    "eslint-plugin-import": "^1.8.0",
+    "eslint-plugin-mocha": "^2.2.0",
     "microtime": "^2.0.0",
     "mocha": "^2.3.4",
-    "semantic-release": "^4.3.5"
+    "semantic-release": "^6.0.0"
   },
   "dependencies": {
     "amqp-coffee": "^0.1.27",
@@ -46,7 +45,7 @@
     "common-errors": "^0.5.3",
     "eventemitter3": "^1.1.1",
     "json-stringify-safe": "^5.0.1",
-    "ms-validation": "^1.0.1",
+    "ms-validation": "^2.0.0",
     "node-uuid": "^1.4.7"
   },
   "release": {
