@@ -1,8 +1,14 @@
+/* eslint-disable no-console, max-len */
+
 const chai = require('chai');
 const expect = chai.expect;
 const Errors = require('common-errors');
 const Promise = require('bluebird');
 const ld = require('lodash');
+const latency = time => {
+  const execTime = process.hrtime(time);
+  return execTime[0] * 1000 + (+(execTime[1] / 1000000).toFixed(3));
+};
 
 describe('AMQPTransport', function AMQPTransportTestSuite() {
   // require module
@@ -68,8 +74,8 @@ describe('AMQPTransport', function AMQPTransportTestSuite() {
     };
 
     return AMQPTransport
-      .connect(opts, (message, headers, actions, callback) => {
-        callback(null, `${message}-response`);
+      .connect(opts, function listener(message, headers, actions, callback) {
+        callback(null, { resp: `${message}-response`, time: process.hrtime() });
       })
       .then(amqp => {
         expect(amqp._amqp.state).to.be.eq('open');
@@ -77,11 +83,32 @@ describe('AMQPTransport', function AMQPTransportTestSuite() {
       });
   });
 
-  it('is able to publish to route consumer', () => (
-    this.amqp.publishAndWait('test.default', 'test-message').then(response => {
-      expect(response).to.be.eq('test-message-response');
-    })
-  ));
+  it('is able to publish to route consumer', () => {
+    const sending = process.hrtime();
+    return this.amqp.publishAndWait('test.default', 'test-message').then((response) => {
+      expect(response.resp).to.be.eq('test-message-response');
+      console.log(response.time);
+      console.log('response sent and received after %s, total time %s', latency(response.time), latency(sending));
+    });
+  });
+
+  it('is able to publish to route consumer:2', () => {
+    const sending = process.hrtime();
+    return this.amqp.publishAndWait('test.default', 'test-message').then((response) => {
+      expect(response.resp).to.be.eq('test-message-response');
+      console.log(response.time);
+      console.log('response sent and received after %s, total time %s', latency(response.time), latency(sending));
+    });
+  });
+
+  it('is able to publish to route consumer:2', () => {
+    const sending = process.hrtime();
+    return this.amqp.publishAndWait('test.default', 'test-message').then((response) => {
+      expect(response.resp).to.be.eq('test-message-response');
+      console.log(response.time);
+      console.log('response sent and received after %s, total time %s', latency(response.time), latency(sending));
+    });
+  });
 
   it('is able to send messages directly to a queue', () => {
     const privateQueue = this.amqp._replyTo;
